move ConfigError ADT members into companion
