+++ core/shared/src/main/scala/laika/ast/RewriteRules.scala
@@ -19,7 +19,7 @@ package laika.ast
 import cats.syntax.all.*
 import laika.ast.RewriteRules.{ ChainedRewriteRules, RewritePhaseBuilder, RewriteRulesBuilder }
 import laika.config.Config.ConfigResult
-import laika.config.ConfigErrors
+import laika.config.ConfigError.ConfigErrors
 import laika.factory.{ RenderFormat, TwoPhaseRenderFormat }
 import laika.rewrite.{ OutputContext, TemplateFormatter, UnresolvedNodeDetector }
 import laika.rewrite.link.LinkResolver
