+++ core/shared/src/test/scala/laika/directive/std/TemplateParserSetup.scala
@@ -16,11 +16,12 @@
 
 package laika.directive.std
 
-import cats.syntax.all._
+import cats.syntax.all.*
 import laika.api.builder.OperationConfig
 import laika.ast.Path.Root
-import laika.ast._
-import laika.config.{ Config, ConfigParser, ValidationError }
+import laika.ast.*
+import laika.config.{ Config, ConfigParser }
+import laika.config.ConfigError.ValidationError
 import laika.directive.DirectiveSupport
 import laika.format.HTML
 import laika.parse.SourceCursor
