+++ io/src/main/scala/laika/format/EPUB.scala
@@ -21,6 +21,7 @@ import laika.api.builder.OperationConfig
 import laika.ast.Path.Root
 import laika.ast.*
 import laika.config.*
+import laika.config.ConfigError.ValidationError
 import laika.factory.*
 import laika.io.model.{ BinaryOutput, RenderedTreeRoot }
 import laika.render.epub.{ ContainerWriter, XHTMLRenderer }
