+++ io/src/main/scala/laika/io/config/ResourceLoader.scala
@@ -17,11 +17,11 @@
 package laika.io.config
 
 import cats.effect.{ Async, Sync }
-import cats.syntax.all._
+import cats.syntax.all.*
 import fs2.io.file.Files
 import laika.ast.DocumentType
 import laika.ast.Path.Root
-import laika.config.ConfigResourceError
+import laika.config.ConfigError.ConfigResourceError
 import laika.io.model.{ FilePath, TextInput }
 
 import java.io.{ FileNotFoundException, InputStream }
