+++ io/src/main/scala/laika/theme/config/Font.scala
@@ -19,6 +19,7 @@ package laika.theme.config
 import laika.ast.Path.Root
 import laika.ast.{ Path, VirtualPath }
 import laika.config.*
+import laika.config.ConfigError.DecodingError
 import laika.io.model.FilePath
 
 /** Represents a font resource, either based on a local classpath or file system resource,
